added views and migrating some commands
every table now has an associated view with a _v prefix
this view has a lot more columns and contains all values that were
originally part of the tables before the schema redesign. It also
contains convenience fields with dynamically computed data like if an
emoji is extern to this guild.

The lib file was removed to prevent the added helper in bin from pulling
in all files of this project. It is bad enough, that it needs all
dependencies.

The kowalski-setup-db helper can be used to migrate a database without
starting the full bot or installing the sqlx programm.