added new migration system and begin role tracking
This migration system allows for much more capable migrations because it
enables to depend on more than the db state for migrations.
An example for when that is useful is the new role tracking system. This
enables us to only consider those user roles, that originate from this
bot. This simplifies reaction roles and score roles and enables more
robust reaction role membership limit tracking and should prevent leaked slots.